refactor. api 목적에 맞게 분리 작성
