+++ src/query/useGetAdListQuery.ts
@@ -1,7 +1,7 @@
 import type { AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getAdList } from '../shared/api/apis';
 import type { IAdResponse } from '../shared/types/types';
+import { getAdList } from './../shared/api/productApi';
 
 const useGetAdListQuery = (keyword: string, page: number) => {
   return useQuery<AxiosResponse<IAdResponse>, Error>(
