+++ src/query/useGetDetailQuery.ts
@@ -1,7 +1,7 @@
 import type { AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getDetail } from '../shared/api/apis';
 import type { IDetailResponse } from '../shared/types/types';
+import { getDetail } from './../shared/api/productApi';
 
 const useGetDetailQuery = (productNo: number, keyword?: string) => {
   return useQuery<AxiosResponse<IDetailResponse>, Error>(
