+++ src/query/useGetMainQuery.ts
@@ -1,7 +1,7 @@
 import type { AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getMain } from '../shared/api/apis';
 import type { TMainResponse } from '../shared/types/types';
+import { getMain } from './../shared/api/productApi';
 
 const useGetMainQuery = () => {
   return useQuery<AxiosResponse<TMainResponse>, Error>('main', () => getMain());
