+++ src/query/useGetOrderDetail.ts
@@ -1,7 +1,7 @@
 import type { AxiosError, AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getOrderDetail } from '../shared/api/apis';
 import type { IMyAPIResOrderDetail } from '../shared/types/types';
+import { getOrderDetail } from './../shared/api/productApi';
 
 export default function useGetOrderDetail(id: string) {
   const order = useQuery<AxiosResponse<IMyAPIResOrderDetail>, AxiosError>(
