+++ src/query/useGetReviewsQuery.ts
@@ -1,7 +1,7 @@
 import type { AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getReviews } from '../shared/api/apis';
 import type { IReviewsResponse } from '../shared/types/types';
+import { getReviews } from './../shared/api/productApi';
 
 const useGetReviewsQuery = (productNo: number, page: number) => {
   return useQuery<AxiosResponse<IReviewsResponse>, Error>(
